+++ addons/backpacks/uiDefines.hpp
@@ -1,12 +1,15 @@
-class RscBackPicture;
+#include "\z\tfar\addons\core\defines.hpp"
+
+class RscBackPicture; 
 class RscEditLCD;
 class HiddenButton;
 class HiddenRotator;
 class HiddenFlip;
 
+#define TF_IDD_BACKGROUND	67676
 //--- rt1523g_radio_dialog
 #define IDD_RT1523G_RADIO_DIALOG		1666
-#define IDC_RT1523G_BACKGROUND	IDD_BACKGROUND
+#define IDC_RT1523G_BACKGROUND	TF_IDD_BACKGROUND
 #define IDC_RT1523G_EDIT		LR_EDIT
 #define IDC_RT1523G_CHANNEL_EDIT	LR_CHANNEL
 #define IDC_RT1523G_ENTER		1610
@@ -31,7 +34,7 @@ class HiddenFlip;
 
 //--- mr3000
 #define IDD_MR3000_RADIO_DIALOG			1998
-#define IDC_MR3000_BACKGROUND		IDD_BACKGROUND
+#define IDC_MR3000_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_MR3000_EDIT			LR_EDIT
 #define IDC_MR3000_CHANNEL_EDIT		LR_CHANNEL
 #define IDC_MR3000_ENTER			2392
@@ -58,7 +61,7 @@ class HiddenFlip;
 
 //--- anprc155
 #define IDD_ANPRC155_RADIO_DIALOG		7666
-#define IDC_ANPRC155_BACKGROUND		IDD_BACKGROUND
+#define IDC_ANPRC155_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_ANPRC155_CHANNEL_EDIT	LR_CHANNEL
 #define IDC_ANPRC155_EDIT		LR_EDIT
 #define IDC_ANPRC155_ENTER		3606
@@ -73,7 +76,7 @@ class HiddenFlip;
 
 //--- anprc210
 #define IDD_ANPRC210_RADIO_DIALOG 8423
-#define IDC_ANPRC210_BACKGROUND		IDD_BACKGROUND
+#define IDC_ANPRC210_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_ANPRC210_CHANNEL_EDIT	1411
 #define IDC_ANPRC210_EDIT		1410
 #define IDC_ANPRC210_ENTER		3552
@@ -96,7 +99,7 @@ class HiddenFlip;
 
 //--- mr6000l_radio_dialog
 #define IDD_MR6000L_RADIO_DIALOG 20135
-#define IDC_MR6000L_BACKGROUND		IDD_BACKGROUND
+#define IDC_MR6000L_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_MR6000L_CHANNEL_EDIT	LR_CHANNEL
 #define IDC_MR6000L_EDIT		LR_EDIT
 #define IDC_MR6000L_ENTER		20536
@@ -120,7 +123,7 @@ class HiddenFlip;
 
 //--- anarc164
 #define IDD_ANARC164_RADIO_DIALOG 3174
-#define IDC_ANARC164_BACKGROUND		IDD_BACKGROUND
+#define IDC_ANARC164_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_ANARC164_CHANNEL_EDIT	LR_CHANNEL
 #define IDC_ANARC164_EDIT		LR_EDIT
 #define IDC_ANARC164_ENTER		3575
@@ -135,7 +138,7 @@ class HiddenFlip;
 
 //--- bussole
 #define IDD_BUSSOLE_RADIO_DIALOG		8666
-#define IDC_BUSSOLE_BACKGROUND		IDD_BACKGROUND
+#define IDC_BUSSOLE_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_BUSSOLE_CHANNEL_EDIT	LR_CHANNEL
 #define IDC_BUSSOLE_EDIT		LR_EDIT
 #define IDC_BUSSOLE_ENTER		8667
