+++ addons/core/functions/fnc_ClientInit.sqf
@@ -168,53 +168,6 @@ TF_last_lr_tangent_press = 0.0;
 TF_last_dd_tangent_press = 0.0;
 
 TF_HintFnc = nil;
-/*
-
-IDC_ANPRC152_RADIO_DIALOG_EDIT_ID = IDC_ANPRC152_EDIT;
-IDC_ANPRC152_RADIO_DIALOG_ID = IDD_ANPRC152_RADIO_DIALOG;
-
-IDC_ANPRC155_RADIO_DIALOG_EDIT_ID = IDC_ANPRC155_EDIT;
-IDC_ANPRC155_RADIO_DIALOG_ID = IDD_ANPRC155_RADIO_DIALOG;
-
-IDC_PRN1000A_RADIO_DIALOG_ID = IDC_PNR1000A_RADIO_DIALOG;
-
-IDC_RF7800STR_RADIO_DIALOG_ID =IDD_RF7800STR_RADIO_DIALOG;
-
-IDC_ANPRC148JEM_RADIO_DIALOG_EDIT_ID = IDC_ANPRC148JEM_EDIT;
-IDC_ANPRC148JEM_RADIO_DIALOG_ID = IDD_ANPRC148JEM_RADIO_DIALOG;
-
-IDC_FADAK_RADIO_DIALOG_EDIT_ID = IDC_FADAK_EDIT;
-IDC_FADAK_RADIO_DIALOG_ID = IDD_FADAK_RADIO_DIALOG;
-
-IDC_RT1523G_RADIO_DIALOG_EDIT_ID = IDC_RT1523G_EDIT;
-IDC_RT1523G_RADIO_DIALOG_ID = IDD_RT1523G_RADIO_DIALOG;
-
-IDC_MR3000_RADIO_DIALOG_EDIT_ID = IDC_MR3000_EDIT;
-IDC_MR3000_RADIO_DIALOG_ID = IDD_MR3000_RADIO_DIALOG;
-
-IDC_MR6000L_RADIO_DIALOG_EDIT_ID = IDC_MR6000L_EDIT;
-IDC_MR6000L_RADIO_DIALOG_ID = IDD_MR6000L_RADIO_DIALOG;
-
-IDC_ANARC164_RADIO_DIALOG_EDIT_ID = IDC_ANARC164_EDIT;
-IDC_ANARC164_RADIO_DIALOG_ID = IDD_ANARC164_RADIO_DIALOG;
-
-IDC_ANPRC210_RADIO_DIALOG_EDIT_ID = IDC_ANPRC210_EDIT;
-IDC_ANPRC210_RADIO_DIALOG_ID = IDD_ANPRC210_RADIO_DIALOG;
-
-IDC_BUSSOLE_RADIO_DIALOG_EDIT_ID = IDC_BUSSOLE_EDIT;
-IDC_BUSSOLE_RADIO_DIALOG_ID = IDD_BUSSOLE_RADIO_DIALOG;
-
-IDC_DIDER_RADIO_DIALOG_ID = IDD_DIVER_RADIO_DIALOG;
-IDC_DIVER_RADIO_EDIT_ID = IDC_DIVER_RADIO_EDIT;
-IDC_DIVER_RADIO_DEPTH_ID = IDC_DIVER_RADIO_DEPTH_EDIT;
-
-TF_BACKGROUND_ID = IDD_BACKGROUND;
-TF_MICRODAGR_BACKGROUND_ID = IDC_MICRODAGR_BACKGROUND;
-TF_MICRODAGR_CLEAR_ID = IDC_MICRODAGR_CLEAR;
-TF_MICRODAGR_ENTER_ID = IDC_MICRODAGR_ENTER;
-TF_MICRODAGR_EDIT_ID = IDC_MICRODAGR_EDIT;
-TF_MICRODAGR_CHANNEL_EDIT_ID = IDC_MICRODAGR_CHANNEL_EDIT;
-*/
 
 TF_tangent_sw_pressed = false;
 TF_tangent_lr_pressed = false;
@@ -398,5 +351,6 @@ diag_log "arsenalClosedHandler";
 }] call bis_fnc_addScriptedEventhandler;
 
 
-
+//Macro to missionVariable sessionTracker needs that. Don't ask me why
+missionNamespace setVariable ["TF_ADDON_VERSION",TF_ADDON_VERSION];
 call TFAR_fnc_sessionTracker;
