--- addons/core/functions/fnc_updateProgrammatorDialog.sqf
@@ -1,9 +1,0 @@
-#include "script_component.hpp"
-
-if !(call TFAR_fnc_haveProgrammator) then {
-    ((_this select 0) displayCtrl TF_MICRODAGR_BACKGROUND_ID) ctrlShow false;
-    ((_this select 0) displayCtrl TF_MICRODAGR_CLEAR_ID) ctrlShow false;
-    ((_this select 0) displayCtrl TF_MICRODAGR_ENTER_ID) ctrlShow false;
-    ((_this select 0) displayCtrl TF_MICRODAGR_EDIT_ID) ctrlShow false;
-    ((_this select 0) displayCtrl TF_MICRODAGR_CHANNEL_EDIT_ID) ctrlShow false;
-};
