+++ addons/handhelds/XEH_preStart.sqf
@@ -1,0 +1,3 @@
+#include "script_component.hpp"
+
+PREP(updateProgrammatorDialog);
