+++ addons/handhelds/uiDefines.hpp
@@ -1,14 +1,10 @@
-class RscBackPicture;
-class RscEditLCD;
-class HiddenButton;
-class HiddenRotator;
-class HiddenFlip;
+#include "\z\tfar\addons\core\defines.hpp"
 
+#define TF_IDD_BACKGROUND	67676
 //--- anprc152_radio_dialog
 #define IDD_ANPRC152_RADIO_DIALOG		1333
-#define IDD_BACKGROUND	67676
 
-#define IDC_ANPRC152_BACKGROUND	IDD_BACKGROUND
+#define IDC_ANPRC152_BACKGROUND	TF_IDD_BACKGROUND
 #define IDC_ANPRC152_EDIT		SW_EDIT
 #define IDC_ANPRC152_ENTER		1600
 #define IDC_ANPRC152_CLEAR		1601
@@ -24,7 +20,7 @@ class HiddenFlip;
 
 //--- anprc148jem
 #define IDD_ANPRC148JEM_RADIO_DIALOG			6000
-#define IDC_ANPRC148JEM_BACKGROUND		IDD_BACKGROUND
+#define IDC_ANPRC148JEM_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_ANPRC148JEM_EDIT			SW_EDIT
 #define IDC_ANPRC148JEM_CHANNEL_EDIT		SW_CHANNEL
 #define IDC_ANPRC148JEM_ENTER		6468
@@ -40,7 +36,7 @@ class HiddenFlip;
 
 //--- fadak
 #define IDD_FADAK_RADIO_DIALOG			7777
-#define IDC_FADAK_BACKGROUND		IDD_BACKGROUND
+#define IDC_FADAK_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_FADAK_EDIT		SW_EDIT
 #define IDC_FADAK_CHANNEL_EDIT	SW_CHANNEL
 #define IDC_FADAK_ENTER_FADAK		2616
@@ -57,7 +53,7 @@ class HiddenFlip;
 //--- diver_radio
 #define IDD_DIVER_RADIO_DIALOG	6900
 
-#define IDC_DIVER_RADIO_BACKGROUND	IDD_BACKGROUND
+#define IDC_DIVER_RADIO_BACKGROUND	TF_IDD_BACKGROUND
 #define IDC_DIVER_RADIO_ENTER	7193
 #define IDC_DIVER_RADIO_CLEAR	7194
 #define IDC_DIVER_RADIO_EDIT	7393
@@ -67,7 +63,7 @@ class HiddenFlip;
 
 //--- anprc154
 #define IDD_ANPRC154_RADIO_DIALOG       3198
-#define IDC_ANPRC154_BACKGROUND       	IDD_BACKGROUND
+#define IDC_ANPRC154_BACKGROUND       	TF_IDD_BACKGROUND
 #define IDC_ANPRC154_RESET_CHANNEL    	3599
 #define IDC_ANPRC154_PREV_CHANNEL   	3601
 #define IDC_ANPRC154_NEXT_CHANNEL    	3602
@@ -80,7 +76,7 @@ class HiddenFlip;
 
 //--- rf7800str
 #define IDD_RF7800STR_RADIO_DIALOG        4425
-#define IDC_RF7800STR_BACKGROUND        IDD_BACKGROUND
+#define IDC_RF7800STR_BACKGROUND        TF_IDD_BACKGROUND
 #define IDC_RF7800STR_NEXT_CHANNEL_ALT    4826
 #define IDC_RF7800STR_PREV_CHANNEL_ALT    4827
 #define IDC_RF7800STR_PREV_CHANNEL    4828
@@ -93,7 +89,7 @@ class HiddenFlip;
 
 //--- pnr1000a
 #define IDC_PNR1000A_RADIO_DIALOG	3083
-#define IDC_PNR1000A_BACKGROUND		IDD_BACKGROUND
+#define IDC_PNR1000A_BACKGROUND		TF_IDD_BACKGROUND
 #define IDC_PNR1000A_RESET_CHANNEL	3484
 #define IDC_PNR1000A_PREV_CHANNEL	3486
 #define IDC_PNR1000A_NEXT_CHANNEL	3487
